MCP_ToolSuite/requirements.txt
mcp
python-dotenv
google-generativeai
tavily-python
requests
markdown-it-py
matplotlib
google-genai